+++ sdk/src/utils/cbor_types.rs
@@ -11,8 +11,10 @@
 // specific language governing permissions and limitations under
 // each license.
 
-use std::fmt;
+use std::{fmt, ops::Deref};
 
+#[cfg(feature = "json_schema")]
+use schemars::JsonSchema;
 use serde::{
     de::{Deserialize, Deserializer},
     ser::{Serialize, Serializer},
@@ -27,6 +29,7 @@ use serde_cbor::tags::Tagged;
 //
 // https://tools.ietf.org/html/rfc7049#section-2.4.1
 #[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "json_schema", derive(JsonSchema))]
 pub struct DateT(pub String);
 
 impl Serialize for DateT {
@@ -48,6 +51,14 @@ impl<'de> Deserialize<'de> for DateT {
 impl AsRef<str> for DateT {
     fn as_ref(&self) -> &str {
         &self.0
+    }
+}
+
+impl Deref for DateT {
+    type Target = str;
+
+    fn deref(&self) -> &Self::Target {
+        self.0.as_str()
     }
 }